Restructure tests in prep for deno testing
